File tree refactoring
- './config/' -> './conf/'
- './conf/processedDefault.js' -> './conf/processedConfig.js'
- './conf/default.json' -> './conf/servers.json'
- Moved all plugin installation files to `openmctStaticServer/plugins`
- Moved plugin configuration files (dictionaries non user definitions, namespaces, etc)
  to `openmctStaticServer/plugins/conf`
