+++ conf/dictionaryVectorCollectionsTelemetry.json
