+++ openmctStaticServer/index.html
@@ -4,13 +4,13 @@
     <title>iCub Telemetry Visualizer (Open MCT based)</title>
     <script src="openmctdist/openmct.js"></script>
     <script src="lib/http.js"></script>
-    <script src="namespacesNtypes.js"></script>
-    <script src="domainObjTypes.js"></script>
-    <script src="genDictFromWalkingCtrlPortDataStruct.js"></script>
-    <script src="dictionary-plugin.js"></script>
-    <script src="historical-telemetry-plugin.js"></script>
-    <script src="realtime-telemetry-plugin.js"></script>
-    <script src="common/processedDefault.json"></script>
+    <script src="plugins/namespacesNtypes.js"></script>
+    <script src="plugins/conf/domainObjTypes.js"></script>
+    <script src="plugins/genDictFromWalkingCtrlPortDataStruct.js"></script>
+    <script src="plugins/dictionary-plugin.js"></script>
+    <script src="plugins/historical-telemetry-plugin.js"></script>
+    <script src="plugins/realtime-telemetry-plugin.js"></script>
+    <script src="common/processedConfig.json"></script>
     <link rel="stylesheet" href="openmctdist/espressoTheme.css">
 </head>
 <body>
