+++ openmctStaticServer/plugins/conf/dictionaryBaseTemplate.json
