+++ openmctStaticServer/plugins/conf/dictionaryIcubTelemetry.json
