+++ openmctStaticServer/plugins/conf/domainObjTypes.js
