+++ openmctStaticServer/plugins/genDictFromWalkingCtrlPortDataStruct.js
