+++ openmctStaticServer/plugins/namespacesNtypes.js
