+++ openmctStaticServer/plugins/realtime-telemetry-plugin.js
